refactor(#3634): split ticket.rs's inline tests into ticket_tests.rs
The `--lite` file-size ratchet FAILed: `ticket.rs` was ALREADY over the ~800-line
source target on `origin/main` (907), and this issue's `token_in_range` coverage
grew it to 1028. The ratchet is right to fail — the campsite rule is that you do
not grow an over-threshold file.

Split by responsibility rather than reaching for `CQLITE_ALLOW_FILE_GROWTH=1`,
following the convention this crate already uses in three places
(`filter_tests.rs`, `bypass_tests.rs`, `statics_tests.rs`): `#[cfg(test)] #[path
= "ticket_tests.rs"] mod tests;`. `ticket.rs` 1028 -> 449 lines.

Pure code motion: no test body changed. Verified the split dropped nothing by
comparing TOTALS, not just a green tally — 21 `ticket::` tests pass and the suite
is 21 + 370 filtered = 391, identical to the 8 + 383 = 391 before the split. A
span-replacing edit that silently deletes cases reports `failed: 0` over a
shrunken suite, which is why the count is the check.
